VideoCommon: simplify and improve depth handling
Instead of using fixed-function depth and, for oversized depth ranges,
falling back to workarounds that introduce rounding errors, we now
always recalculate the depth value in the pixel shader but only write it
back ("slow depth") when we know the fixed-function depth is inaccurate.
This means the option to always force slow depth is now unnecessary.

Additionally, if the backend supports unrestricted depth range, we don't
need to recalculate the depth value, just clamp it. This means slow
depth is now slightly faster.

+++ Source/Core/VideoCommon/FramebufferManager.h
@@ -103,7 +103,7 @@ public:
 
   // Writes a value to the framebuffer. This will never block, and writes will be batched.
   void PokeEFBColor(u32 x, u32 y, u32 color);
-  void PokeEFBDepth(u32 x, u32 y, float depth);
+  void PokeEFBDepth(u32 x, u32 y, u32 depth);
   void FlushEFBPokes();
 
   // Save state load/save.
